lwip: put the mempools in AHBSRAM1
We don't have enough lower memory, we need to use one of the high
memories for some data. AHBSRAM0 is already used for the LWIP heap.
Put the LWIP mempools in AHBSRAM1.

This is a big hack, because it forces something platform-specific
in the global memp.h. Ideally we would do this outside of mbed-os
entirely, by setting configuration options, e.g. overriding
LWIP_DECLARE_MEMORY_ALIGNED. However, that symbol is also used for
the LWIP heap, which adds another __attribute__((section)) so that
gives a conflict.