DHCP: reduce coarse timer to 15 sec
The VOO router has an extremely short lease time when there is no
internet connection. Because the least time is the same order as the
DHCP coarse timer, instead of renewing the lease, the DHCP client
will release and restart a full DHCP request (to be precise: the t1
and t2 timers are never armed, instead only the t0 timer is armed and
it is that one that times out).

As a quick workaround, reduce the coarse timer from 60 seconds to 15
seconds. This way, a renew is attempted even for a lease time of 30
seconds.